users/requirements.txt
django
djangorestframework
djangorestframework-jsonapi
django-filter
django-tables2